requirement.txt
beautifulsoup4>=4.9
requests>=2.24
python-telegram-bot>=12.5.1
